表單重構 (#890)
* reuse textinput

* reuse textarea

* refactor: text-input use-enter-confirm

* merge auto-complete-text-input & text-input

* move autocomplete-search-text-input

* remove textarea unused props
