+++ src/components/App/Header/Searchbar.js
@@ -4,7 +4,7 @@ import cn from 'classnames';
 import { withRouter } from 'react-router-dom';
 import qs from 'qs';
 
-import AutoCompleteSearchTextInput from 'common/form/AutoCompleteTextInput_new/AutoCompleteSearchTextInput';
+import SearchTextInput from 'common/form/TextInput/SearchTextInput';
 import Magnifier from '../../common/icons/Magnifiner';
 import styles from './Searchbar.module.css';
 
@@ -47,7 +47,7 @@ const Searchbar = ({ className, placeholder, history, location }) => {
       onFocus={handleFormFocus}
       onBlur={handleFormBlur}
     >
-      <AutoCompleteSearchTextInput
+      <SearchTextInput
         className={styles.textInput}
         placeholder={placeholder}
         value={searchText}
