+++ src/components/common/FormBuilder/QuestionBuilder/Text.js
@@ -4,8 +4,7 @@ import PropTypes from 'prop-types';
 import cn from 'classnames';
 import R from 'ramda';
 
-import AutoCompleteTextInput from 'common/form/AutoCompleteTextInput_new';
-import useComposition from '../useComposition';
+import TextInput from 'common/form/TextInput';
 import styles from './Text.module.css';
 
 const notEquals = x =>
@@ -30,11 +29,6 @@ const Text = ({
   validator,
   placeholder,
 }) => {
-  const [
-    isComposing,
-    handleCompositionStart,
-    handleCompositionEnd,
-  ] = useComposition();
   const [items, setItems] = useState([]);
   const ref = useRef(null);
 
@@ -61,20 +55,16 @@ const Text = ({
 
   return (
     <div className={cn(styles.container, { [styles.hasWarning]: !!warning })}>
-      <AutoCompleteTextInput
+      <TextInput
         ref={ref}
         className={styles.textinput}
         type="text"
         placeholder={placeholder}
         value={value}
-        onCompositionStart={handleCompositionStart}
-        onCompositionEnd={handleCompositionEnd}
         onChange={e => onChange(e.target.value)}
         onEnter={e => {
-          if (!isComposing) {
-            e.target.blur();
-            onConfirm(e);
-          }
+          e.target.blur();
+          onConfirm(e);
         }}
         autocompleteItems={items}
         onAutocompleteItemSelected={item => {
