--- src/components/common/FormBuilder/useComposition.js
@@ -1,8 +1,0 @@
-import { useState } from 'react';
-
-export default () => {
-  const [isComposing, setComposing] = useState(false);
-  const handleCompositionStart = () => setComposing(true);
-  const handleCompositionEnd = () => setComposing(false);
-  return [isComposing, handleCompositionStart, handleCompositionEnd];
-};
