--- src/components/common/form/AutoCompleteTextInput_new/styles.module.css
@@ -1,38 +1,0 @@
-.textInput {
-  width: 100%;
-}
-
-.menu {
-  display: none;
-  border: 1px solid #b4b4b4;
-  max-height: 150px;
-  overflow-y: scroll;
-  background: #fff;
-
-  &.visible {
-    display: block;
-  }
-}
-
-.item {
-  background: white;
-  padding: 5px 10px;
-  color: #222;
-
-  &.active,
-  &:hover {
-    background: lightgray;
-    color: black;
-  }
-}
-
-.wrapper {
-  position: relative;
-
-  .menu {
-    position: absolute;
-    left: 0;
-    right: 0;
-    z-index: 99;
-  }
-}
